requirements.txt
# Core Libraries
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
nltk>=3.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0

# Elasticsearch
elasticsearch>=8.0.0,<9.0.0 # Or version matching your cluster

# Text Processing & Sentiment
afinn>=0.1 # Added dependency

# BERT & Deep Learning
tensorflow>=2.8.0 # Or torch if using PyTorch version of transformers
transformers>=4.15.0
datasets # Often used with transformers, e.g., for IMDB training dataset

# Image Processing & OCR
opencv-python-headless>=4.5.0 # Use headless if no GUI needed
pytesseract>=0.3.8
Pillow>=9.0.0

# Jupyter (Optional - if running notebooks)
# jupyterlab>=3.0.0
# notebook>=6.0.0

# Other utilities potentially used (review if needed)
# tqdm # Progress bars
# enchant # Dictionary checking (if used)

# NOTE: Review versions if specific compatibility is needed.
# NOTE: Ensure system dependencies like Tesseract OCR are installed.